+++ Source/Lib/TraktApiSharp/Objects/Basic/TraktRating.cs
@@ -6,7 +6,7 @@
     public class TraktRating
     {
         [JsonProperty(PropertyName = "rating")]
-        public decimal? Rating { get; set; }
+        public float? Rating { get; set; }
 
         [JsonProperty(PropertyName = "votes")]
         public int? Votes { get; set; }
